Create Nov10.java
MY OPINION --> 

This is question number 27 in LeetCode.
Its difficulty is easy though, anyone who knows the basics can easily solve this problem.

QUESTION DESCRIPTION -->
 
Basically, this question states that you are given with an array named 'nums' and you have to remove a given element named 'val' from it and return the length of the newly formed array from it.

MY APPROACH --> 

So, I have used a very basic approach of solving this question is to create a new temporary array named 'expectedNums' which will store those elements which are not equal to given value to remove by iterating through the elements one by one and checking if the current value is not equal to 'val'. If this condition is true then we will store this current element in our new array and finally return the length of our new array.

TIME COMPLEXITY --> O (n)
Because we are iterating through the whole array.

I hope this will help you in understanding.