Create Nov11.java
QUESTION DESCRIPTION --> 

Question 35: Search Insert position.
This question is of binary search in which we have to search a given element in a sorted array of length 'n' in O (log n) time complexity.

APPROACH -->

I have used the iterative way to solve this question instead of recursive way. I know both are valid but since iterative way is a little bit faster than recursive way, that's why I chose to solve it by while loop in which we create two pointers left and right, initially set to 0 and nums.length -1. 
If middle element is equal to our target, then we can simply return the middle element, if not then we check if its greater or smaller than the middle element.
If it's greater then, we check for half the portion on the right-hand side, if not then we check for half the portion on the left-hand side until and unless the left pointer becomes greater than the right pointer.
And, if the search is unsuccessful then we return the left pointer itself where we have to insert it.