Create Nov13.java
MY OPINION --> 

Question 700: Search in a BST
It is a very simple question of Binary search tree in which we have to find a given element in a BST.

APPROACH --> 

First of all, if the root is null then we can simply return null as there is no value in BST.
If the given value is equal to the root value, then we can simply return root value.
If root value is greater than the given value, then we can simply search for the element in LHS otherwise we can search for the given element in RHS with the help of recursion.
We don't have to go deep into the recursive part as it may become complicated to understand what's going on. Hence, we should have a leap of faith in recursion that it will solve the problem on its own.