Create Nov14.java
MY OPINION --> 

Question 1: Two Sum.
This question seems easy and is easy too. Because the first thought after seeing this question comes to our mind that we can use two for loops to search the two elements and then return their indices in the form of an array. But it will take some time to do so. 
So, what we can do is that we can make use of some space to lessen our time complexity. 

QUESTION DESCRIPTION --> 

You have to find two elements in an array which will sum up to the given target and return their indices in the form of an array.
If there are no two elements found, then return an empty array.

APPROACH -->
 
I have taken help of hash map in which the keys are nums elements and values are their indices. Now we use a for loop to search if the complement pair is present in the hash map or not.
For eg. 
The array is [2,5,9,11] and the target is 7. 
So, the answer is [0,1] since the values at indices 0 and 1 are 2 and 5 which will sum up to give the target 7.
Now let's take a look at our for loop.
For i = 0, the complement is7 - 2 = 5 and 5 is not present in our map, so we put 2 ,0 in our hash map.
For i = 1, the complement is 7-5 = 2 and 2 is present in our hash map.
And as soon as we get our complement in our hash map we return the array from there only as [map.get(complement) , i] = [map.get(2) , 1] = [0,1].
And there you go our answer.