Create Nov15.java
MY OPINION --> 

Question 11: Container with most water.
This is a medium level question which uses the concept used in DSA known as two pointers.
Let me Explain.....

QUESTION DESCRIPTION -->

The question states that you are given an array of height which stores the height of individual pillars.
And you have to find the max. area of container with most water.
For eg.
arr[] = {1,2,3,4,5}
So, it can be drawn like this -->
        |
      | |
    | | |
  | | | |
| | | | |
The container with most water is either 1*4 or 4*1 = 4.

APPROACH -->

So, we firstly create two pointers left and right initially stored from 0 to length -1.
Now, we iterate until the left pointer becomes greater than the right pointer and initialize a variable area = 0.
Now, in each iteration we have to see the maximum area that is coming from our two pointers each time.
And also, don't forget to increment and decrement the pointers.
If height of left pointer is greater than right pointer, we move the left pointer by one to the right and else, we move the right pointer by decrementing it by one (means moving it to the left).
right - left pointer will give us the width for calculating the area.