Create Nov20.java
MY OPININON --> 
 
Question 172: Trailing zeroes in n!
In this question, we have to find the number of trailing zeroes in n!. 
One can think of finding n! first and then counting the trailing zeroes by %10. But that would take so long.

APPROACH --> 

Trailing zeroes in a number result from multiplying factors of 10. Since 10=2×5, the number of trailing zeroes is determined by the number of pairs of factors of 2 and 5 in n!.

In factorials, there are always more factors of 2 than 5, so the number of trailing zeroes is determined by the count of 5s in the prime factorization of n!.

To count the factors of 5, we count how many multiples of 5, 25, 125, etc., are less than or equal to n. Each contributes additional 5s to the factorization.

Initialize a counter count to store the number of trailing zeroes.
Start with divisor=5 and repeatedly divide n by divisor, adding the quotient to count.
Continue until divisor>n.